Advanced Identify Widget
The organization I work at (NASUVINSA/LURSAREA) is developing the cAAE
project (Industry Catalogue of Navarre, Spain) in collaboration with the
regional government. As part of the products of the project we are
publishing a web app developed with Web AppBuilder for ArcGIS. The
public app allows to search for availability of industrial land as well
as the current economic activities in the region. In order to improve
user experience, we created together with ESRI Spain a custom widget
(Advanced Identify) that allows to organize layer content in a more
intuitive way. Feel free to test and use in your own projects. Hope you
will find it useful.